fix o_ret tyep
git-svn-id: http://code.taobao.org/svn/tfs@608 b722c2b5-6a69-4643-bcd1-1a75825fa998

+++ branches/tfs-dev-1.4.1/src/name_meta_server/data_store_mysql/create_file.sql
@@ -6,7 +6,7 @@ create_file(in i_app_id bigint, in i_uid bigint,
   in i_name varbinary(512))
 begin
     declare aff_row int;
-    declare o_ret int;
+    declare o_ret bigint;
     declare exit handler for sqlexception
     begin
         set o_ret = -1;
