+++ branches/tfs-dev-1.4.1/src/name_meta_server/data_store_mysql/rm_file.sql
@@ -5,7 +5,7 @@ rm_file(in i_app_id bigint, in i_uid bigint, in i_ppid bigint unsigned,
   in i_pid bigint unsigned, in i_pname varbinary(512), in i_name varbinary(512))
 begin
   declare aff_row int;
-  declare o_ret int;
+  declare o_ret bigint;
   declare name_end varbinary(512);
   declare exit handler for sqlexception
   begin
